% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{canBeNumeric}
\alias{canBeNumeric}
\title{Determines whether an object is or can be converted to numeric}
\usage{
canBeNumeric(object)
}
\arguments{
\item{object}{the object}
}
\description{
Determines whether an object is or can be converted to numeric
}
