% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themes.R
\name{colorPalette}
\alias{colorPalette}
\title{A function that creates a color palette}
\usage{
colorPalette(n = 5, pal = "jmv", type = "fill")
}
\arguments{
\item{n}{Number of colors needed}

\item{pal}{Color palette name}

\item{type}{'fill' or 'color'}
}
\value{
a vector of hex color codes
}
\description{
A function that creates a color palette
}
