% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{composeFormula}
\alias{composeFormula}
\title{Compose a formula string}
\usage{
composeFormula(lht, rht)
}
\arguments{
\item{lht}{list of character vectors making up the left}

\item{rht}{list of character vectors making up the right}
}
\value{
a string representation of the formula
}
\description{
Compose a formula string
}
\examples{

composeFormula(list('a', 'b', c('a', 'b')))
# ~a+b+a:b

composeFormula('f', list('a', 'b', c('a', 'b')))
# "f~a+b+a:b"

composeFormula('with spaces', list('a', 'b', c('a', 'b')))
'`with spaces`~a+b+a:b'

}
