% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{decomposeFormula}
\alias{decomposeFormula}
\title{Decompose a formula}
\usage{
decomposeFormula(formula)
}
\arguments{
\item{formula}{the formula to decompose}
}
\value{
a list of lists of the formulas components
}
\description{
Decompose a formula
}
