% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{composeTerm}
\alias{composeTerm}
\alias{composeTerms}
\alias{decomposeTerm}
\alias{decomposeTerms}
\title{Compose and decompose interaction terms to and from their components}
\usage{
composeTerm(components)

composeTerms(listOfComponents)

decomposeTerm(term)

decomposeTerms(terms)
}
\arguments{
\item{components}{a character vectors of components}

\item{listOfComponents}{a list of character vectors of components}

\item{term}{a string with components separated with colons}

\item{terms}{a character vector of components separated with colons}
}
\description{
Compose and decompose interaction terms to and from their components
}
\examples{
composeTerm(c('a', 'b', 'c'))
# 'a:b:c'

composeTerm(c('a', 'b', 'with space'))
# 'a:b:`with space`'

decomposeTerm('a:b:c')
# c('a', 'b', 'c')

decomposeTerm('a:b:`with space`')
# c('a', 'b', 'with space')

}
