% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/i18n.R
\name{..}
\alias{..}
\title{Designate string as client-side translated}
\usage{
..(format, values = NULL)
}
\arguments{
\item{format}{a format string}

\item{values}{a string or named list of strings to substitute}
}
\description{
Designate string as client-side translated
}
