% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/i18n.R
\name{.}
\alias{.}
\title{Designate string as translated}
\usage{
.(text, n = 1)
}
\arguments{
\item{text}{the string to translate}

\item{n}{the number (if applicable) for pluralisation}
}
\description{
Designate string as translated
}
