% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{enquo}
\alias{enquo}
\title{rlang::enquo
Simplifies things so packages overriding Analysis don't need
to have rlang in their imports.
This is intended for use by classes overriding Analysis}
\usage{
enquo(arg)
}
\arguments{
\item{arg}{the argument to enquote}
}
\value{
the quosure
}
\description{
rlang::enquo
Simplifies things so packages overriding Analysis don't need
to have rlang in their imports.
This is intended for use by classes overriding Analysis
}
