% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{extractErrorMessage}
\alias{extractErrorMessage}
\title{Extracts the error message from an error object}
\usage{
extractErrorMessage(error)
}
\arguments{
\item{error}{an error object}
}
\description{
Extracts the error message from an error object
}
