% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{htmlEscape}
\alias{htmlEscape}
\title{Escape HTML special characters}
\usage{
htmlEscape(text, attribute = FALSE)
}
\arguments{
\item{text}{a character vector to escape}

\item{attribute}{logical; if TRUE, also escapes quotes and newlines for use
in HTML attribute values}
}
\value{
the escaped character vector
}
\description{
Escape HTML special characters
}
