% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{isError}
\alias{isError}
\title{Determine if an object is an error}
\usage{
isError(object)
}
\arguments{
\item{object}{the object to test}
}
\value{
TRUE if the object is an error
}
\description{
Determine if an object is an error
}
