% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marshal.R
\name{marshalData}
\alias{marshalData}
\title{Marshal the data from an environment into a data frame}
\usage{
marshalData(env, ...)
}
\arguments{
\item{env}{the environment to marshal from}

\item{...}{the variables to marshal}
}
\value{
a data frame
}
\description{
Marshal the data from an environment into a data frame
}
