% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marshal.R
\name{marshalFormula}
\alias{marshalFormula}
\title{Marshal a formula into options}
\usage{
marshalFormula(
  formula,
  data,
  from = "rhs",
  type = "vars",
  permitted = c("numeric", "factor"),
  subset = ":",
  required = FALSE
)
}
\arguments{
\item{formula}{the formula}

\item{data}{a data frame to marshal the data from}

\item{from}{'rhs' or 'lhs', which side of the formula should be marshalled}

\item{type}{'vars' or 'terms', the type of the option be marshalled to}

\item{permitted}{the types of data the option permits}

\item{subset}{a subset of the formula to marshal}

\item{required}{whether this marshall is required or not}
}
\description{
Marshal a formula into options
}
