% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{matchSet}
\alias{matchSet}
\title{Determines the index where an item appears}
\usage{
matchSet(x, table)
}
\arguments{
\item{x}{the item to find}

\item{table}{the object to search}
}
\value{
the index of where the item appears, or -1 if it isn't present
}
\description{
Determines the index where an item appears
}
