% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{naOmit}
\alias{naOmit}
\title{remove missing values from a data frame listwise}
\usage{
naOmit(object)
}
\arguments{
\item{object}{the object to remove missing values from}
}
\description{
removes all rows from the data frame which contain missing values (NA)
}
\details{
this function is equivalent to \code{\link{na.omit}} from the stats package,
however it preserves attributes on columns in data frames
}
