% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\name{Options}
\alias{Options}
\alias{OptionBool}
\alias{OptionAction}
\alias{OptionList}
\alias{OptionNMXList}
\alias{OptionVariables}
\alias{OptionTerm}
\alias{OptionVariable}
\alias{OptionOutput}
\alias{OptionTerms}
\alias{OptionInteger}
\alias{OptionNumber}
\alias{OptionString}
\alias{OptionLevel}
\alias{OptionFile}
\alias{OptionGroup}
\alias{OptionPair}
\alias{OptionSort}
\alias{OptionArray}
\alias{OptionPairs}
\title{The jmv Options classes}
\description{
The jmv Options classes
}
