% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{createError}
\alias{createError}
\alias{reject}
\title{Create and throw errors}
\usage{
createError(formats, code = NULL, ...)

reject(formats, code = NULL, ...)
}
\arguments{
\item{formats}{a format string which is passed to \code{\link{format}}}

\item{code}{an error code}

\item{...}{additional arguments passed to \code{\link{format}}}
}
\description{
These functions are convenience functions for creating and throwing errors.
}
