% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{resolveQuo}
\alias{resolveQuo}
\title{Evaluates a quosure
This is intended for use by classes overriding Analysis}
\usage{
resolveQuo(quo)
}
\arguments{
\item{quo}{the quosure to evaluate}
}
\value{
the value of the quosure
}
\description{
Evaluates a quosure
This is intended for use by classes overriding Analysis
}
