% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{select}
\alias{select}
\title{Create a new data frame with only the selected columns}
\usage{
select(df, columnNames)
}
\arguments{
\item{df}{the data frame}

\item{columnNames}{the names of the columns to make up the new data frame}
}
\value{
the new data frame
}
\description{
Shorthand equivalent to \code{\link{subset}(df, select=columnNames)}, however
it additionally preserves attributes on the columns and the data frame
}
