% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{startsWith}
\alias{startsWith}
\alias{endsWith}
\title{Test whether strings start or end with a particular string}
\usage{
startsWith(x, prefix)

endsWith(x, suffix)
}
\arguments{
\item{x}{a string to test}

\item{prefix}{a string to test the presence of}

\item{suffix}{a string to test the presence of}
}
\description{
Same as \code{base::startsWith()} and \code{base::endsWith()} except
available for R < 3.3
}
