% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themes.R
\name{theme_blackwhite}
\alias{theme_blackwhite}
\title{Creates the black & white jmv ggplot2 theme}
\usage{
theme_blackwhite(base_size = 16, scale = "none", palette = "Greys")
}
\arguments{
\item{base_size}{Font size}

\item{scale}{'none', 'discrete' or 'continuous'}

\item{palette}{Color palette name}
}
\value{
the black & white jmv ggplot2 theme
}
\description{
Creates the black & white jmv ggplot2 theme
}
