% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themes.R
\name{theme_grid}
\alias{theme_grid}
\title{Creates the grid jmv ggplot2 theme}
\usage{
theme_grid(base_size = 16, scale = "none", palette = "jmv")
}
\arguments{
\item{base_size}{Font size}

\item{scale}{'none', 'discrete' or 'continuous'}

\item{palette}{Color palette name}
}
\value{
the grid jmv ggplot2 theme
}
\description{
Creates the grid jmv ggplot2 theme
}
