% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themes.R
\name{theme_jupwr_ciemny}
\alias{theme_jupwr_ciemny}
\title{Creates the jUPWR lecture-note ggplot2 theme (dark)}
\usage{
theme_jupwr_ciemny(base_size = 16, scale = "none", palette = "jupwrCiemny")
}
\arguments{
\item{base_size}{Font size}

\item{scale}{'none', 'discrete' or 'continuous'}

\item{palette}{Color palette name}
}
\value{
the dark lecture-note ggplot2 theme
}
\description{
Creates the jUPWR lecture-note ggplot2 theme (dark)
}
