% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themes.R
\name{theme_jupwr_jasny}
\alias{theme_jupwr_jasny}
\title{Creates the jUPWR lecture-note ggplot2 theme (light)}
\usage{
theme_jupwr_jasny(base_size = 16, scale = "none", palette = "jupwrJasny")
}
\arguments{
\item{base_size}{Font size}

\item{scale}{'none', 'discrete' or 'continuous'}

\item{palette}{Color palette name}
}
\value{
the light lecture-note ggplot2 theme
}
\description{
Creates the jUPWR lecture-note ggplot2 theme (light)
}
