% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themes.R
\name{theme_spss}
\alias{theme_spss}
\title{Creates the spss jmv ggplot2 theme}
\usage{
theme_spss(base_size = 16, scale = "none", palette = "jmv")
}
\arguments{
\item{base_size}{Font size}

\item{scale}{'none' or 'discrete'}

\item{palette}{Color palette name}
}
\value{
the spss jmv ggplot2 theme
}
\description{
Creates the spss jmv ggplot2 theme
}
