% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themes.R
\docType{data}
\name{themes}
\alias{themes}
\title{The jamovi plot themes}
\format{An object of class \code{R6} of length 7.}
\usage{
themes
}
\description{
This contains the following plot themes:
`themes$default` is the default theme
`themes$hadley` is a theme based on the default ggplot2 theme
`themes$iheartspss` is a theme based on the default SPSS theme
`themes$minimal` is a minimal theme
`themes$liberace` is a gold theme
}
\keyword{datasets}
