% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{toB64}
\alias{toB64}
\alias{fromB64}
\title{Convert names to and from Base64 encoding}
\usage{
toB64(names)

fromB64(names)
}
\arguments{
\item{names}{the names to be converted base64}
}
\description{
Note: uses the . and _ characters rather than + and / allowing these to be
used as variable names
}
