% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{toNumeric}
\alias{toNumeric}
\title{Converts a vector of values to numeric}
\usage{
toNumeric(object)
}
\arguments{
\item{object}{the vector to convert}
}
\description{
Similar to \code{\link{as.numeric}}, however if the object has a values
attribute attached, these are used as the numeric values
}
