% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{tryNaN}
\alias{tryNaN}
\title{try an expression, and return NaN on failure}
\usage{
tryNaN(expr)
}
\arguments{
\item{expr}{an expression to evaluate}
}
\value{
the result, or NaN on failure
}
\description{
if the expression fails, NaN is returned silently
}
