% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/action.R, R/analysis.R, R/array.R, R/column.R,
%   R/group.R, R/html.R, R/image.R, R/notice.R, R/output.R, R/preformatted.R,
%   R/state.R, R/svg.R, R/table.R
\docType{data}
\name{Action}
\alias{Action}
\alias{Analysis}
\alias{Array}
\alias{Column}
\alias{Group}
\alias{Html}
\alias{Image}
\alias{Notice}
\alias{Output}
\alias{Preformatted}
\alias{State}
\alias{Svg}
\alias{Table}
\title{the jmvcore Object classes}
\format{An object of class \code{R6ClassGenerator} of length 25.}
\usage{
Action

Analysis

Array

Column

Group

Html

Image

Notice

Output

Preformatted

State

Svg

Table
}
\description{
the jmvcore Object classes
}
\keyword{datasets}
