% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R, R/array.R, R/column.R, R/group.R, R/html.R, R/image.R, R/preformatted.R, R/table.R
\docType{data}
\name{Analysis}
\alias{Analysis}
\alias{Array}
\alias{Column}
\alias{Group}
\alias{Html}
\alias{Image}
\alias{Preformatted}
\alias{Table}
\title{the jmvcore Object classes}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
Analysis

Array

Column

Group

Html

Image

Preformatted

Table
}
\description{
the jmvcore Object classes
}
\keyword{datasets}

