% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cell.R
\docType{data}
\name{Cell.BEGIN_GROUP}
\alias{Cell.BEGIN_GROUP}
\alias{Cell.END_GROUP}
\alias{Cell.BEGIN_END_GROUP}
\alias{Cell.NEGATIVE}
\alias{Cell.INDENTED}
\title{Constants to specify formatting of Table cells}
\format{
An object of class \code{numeric} of length 1.

An object of class \code{numeric} of length 1.

An object of class \code{numeric} of length 1.

An object of class \code{numeric} of length 1.

An object of class \code{numeric} of length 1.
}
\usage{
Cell.BEGIN_GROUP

Cell.END_GROUP

Cell.BEGIN_END_GROUP

Cell.NEGATIVE

Cell.INDENTED
}
\description{
Cell.BEGIN_GROUP adds spacing above a cell
}
\details{
Cell.END_GROUP add spacing below a cell

Cell.BEGIN_END_GROUP add spacing above and below a cell

Cell.NEGATIVE specifies that the cells contents is negative
}
\examples{
\dontrun{

table$addFormat(rowNo=1, col=1, Cell.BEGIN_END_GROUP)
}

}
\keyword{datasets}
