% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/notice.R
\docType{data}
\name{NoticeType}
\alias{NoticeType}
\title{Different notice levels}
\format{An object of class \code{list} of length 4.}
\usage{
NoticeType
}
\description{
Different notice levels
}
\keyword{datasets}
