% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\name{Options}
\alias{Options}
\alias{OptionBool}
\alias{OptionAction}
\alias{OptionList}
\alias{OptionNMXList}
\alias{OptionVariables}
\alias{OptionTerm}
\alias{OptionVariable}
\alias{OptionOutput}
\alias{OptionTerms}
\alias{OptionInteger}
\alias{OptionNumber}
\alias{OptionString}
\alias{OptionLevel}
\alias{OptionGroup}
\alias{OptionPair}
\alias{OptionSort}
\alias{OptionArray}
\alias{OptionPairs}
\title{The jmv Options classes}
\description{
The jmv Options classes

The jmv Options classes
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Options-new}{\code{Options$new()}}
\item \href{#method-Options-.addOption}{\code{Options$.addOption()}}
\item \href{#method-Options-.getData}{\code{Options$.getData()}}
\item \href{#method-Options-check}{\code{Options$check()}}
\item \href{#method-Options-values}{\code{Options$values()}}
\item \href{#method-Options-translate}{\code{Options$translate()}}
\item \href{#method-Options-eval}{\code{Options$eval()}}
\item \href{#method-Options-.eval}{\code{Options$.eval()}}
\item \href{#method-Options-option}{\code{Options$option()}}
\item \href{#method-Options-get}{\code{Options$get()}}
\item \href{#method-Options-has}{\code{Options$has()}}
\item \href{#method-Options-.removeOption}{\code{Options$.removeOption()}}
\item \href{#method-Options-levels}{\code{Options$levels()}}
\item \href{#method-Options-read}{\code{Options$read()}}
\item \href{#method-Options-asProtoBuf}{\code{Options$asProtoBuf()}}
\item \href{#method-Options-fromProtoBuf}{\code{Options$fromProtoBuf()}}
\item \href{#method-Options-compProtoBuf}{\code{Options$compProtoBuf()}}
\item \href{#method-Options-fromJSON}{\code{Options$fromJSON()}}
\item \href{#method-Options-clone}{\code{Options$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Options-new"></a>}}
\if{latex}{\out{\hypertarget{method-Options-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Options$new(package = "jmv", name = "", requiresData = TRUE, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Options-.addOption"></a>}}
\if{latex}{\out{\hypertarget{method-Options-.addOption}{}}}
\subsection{Method \code{.addOption()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Options$.addOption(option)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Options-.getData"></a>}}
\if{latex}{\out{\hypertarget{method-Options-.getData}{}}}
\subsection{Method \code{.getData()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Options$.getData()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Options-check"></a>}}
\if{latex}{\out{\hypertarget{method-Options-check}{}}}
\subsection{Method \code{check()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Options$check(checkValues = FALSE, checkVars = FALSE, checkData = FALSE)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Options-values"></a>}}
\if{latex}{\out{\hypertarget{method-Options-values}{}}}
\subsection{Method \code{values()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Options$values()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Options-translate"></a>}}
\if{latex}{\out{\hypertarget{method-Options-translate}{}}}
\subsection{Method \code{translate()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Options$translate(text, n = 1)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Options-eval"></a>}}
\if{latex}{\out{\hypertarget{method-Options-eval}{}}}
\subsection{Method \code{eval()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Options$eval(value, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Options-.eval"></a>}}
\if{latex}{\out{\hypertarget{method-Options-.eval}{}}}
\subsection{Method \code{.eval()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Options$.eval(text)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Options-option"></a>}}
\if{latex}{\out{\hypertarget{method-Options-option}{}}}
\subsection{Method \code{option()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Options$option(name)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Options-get"></a>}}
\if{latex}{\out{\hypertarget{method-Options-get}{}}}
\subsection{Method \code{get()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Options$get(name)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Options-has"></a>}}
\if{latex}{\out{\hypertarget{method-Options-has}{}}}
\subsection{Method \code{has()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Options$has(name)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Options-.removeOption"></a>}}
\if{latex}{\out{\hypertarget{method-Options-.removeOption}{}}}
\subsection{Method \code{.removeOption()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Options$.removeOption(name)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Options-levels"></a>}}
\if{latex}{\out{\hypertarget{method-Options-levels}{}}}
\subsection{Method \code{levels()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Options$levels(x)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Options-read"></a>}}
\if{latex}{\out{\hypertarget{method-Options-read}{}}}
\subsection{Method \code{read()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Options$read(raw)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Options-asProtoBuf"></a>}}
\if{latex}{\out{\hypertarget{method-Options-asProtoBuf}{}}}
\subsection{Method \code{asProtoBuf()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Options$asProtoBuf()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Options-fromProtoBuf"></a>}}
\if{latex}{\out{\hypertarget{method-Options-fromProtoBuf}{}}}
\subsection{Method \code{fromProtoBuf()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Options$fromProtoBuf(pb)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Options-compProtoBuf"></a>}}
\if{latex}{\out{\hypertarget{method-Options-compProtoBuf}{}}}
\subsection{Method \code{compProtoBuf()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Options$compProtoBuf(pb)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Options-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-Options-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Options$fromJSON(json)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Options-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Options-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Options$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
\section{Super class}{
\code{jmvcore::Option} -> \code{OptionBool}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-OptionBool-new}{\code{OptionBool$new()}}
\item \href{#method-OptionBool-clone}{\code{OptionBool$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="jmvcore" data-topic="Option" data-id=".getData"><a href='../../jmvcore/html/Option.html#method-Option-.getData'><code>jmvcore::Option$.getData()</code></a></span></li>
<li><span class="pkg-link" data-pkg="jmvcore" data-topic="Option" data-id=".setParent"><a href='../../jmvcore/html/Option.html#method-Option-.setParent'><code>jmvcore::Option$.setParent()</code></a></span></li>
<li><span class="pkg-link" data-pkg="jmvcore" data-topic="Option" data-id="check"><a href='../../jmvcore/html/Option.html#method-Option-check'><code>jmvcore::Option$check()</code></a></span></li>
<li><span class="pkg-link" data-pkg="jmvcore" data-topic="Option" data-id="getBoundValue"><a href='../../jmvcore/html/Option.html#method-Option-getBoundValue'><code>jmvcore::Option$getBoundValue()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OptionBool-new"></a>}}
\if{latex}{\out{\hypertarget{method-OptionBool-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OptionBool$new(name, value = FALSE, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OptionBool-clone"></a>}}
\if{latex}{\out{\hypertarget{method-OptionBool-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OptionBool$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
\section{Super classes}{
\code{jmvcore::Option} -> \code{\link[jmvcore:OptionBool]{jmvcore::OptionBool}} -> \code{OptionAction}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-OptionAction-clone}{\code{OptionAction$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="jmvcore" data-topic="Option" data-id=".getData"><a href='../../jmvcore/html/Option.html#method-Option-.getData'><code>jmvcore::Option$.getData()</code></a></span></li>
<li><span class="pkg-link" data-pkg="jmvcore" data-topic="Option" data-id=".setParent"><a href='../../jmvcore/html/Option.html#method-Option-.setParent'><code>jmvcore::Option$.setParent()</code></a></span></li>
<li><span class="pkg-link" data-pkg="jmvcore" data-topic="Option" data-id="check"><a href='../../jmvcore/html/Option.html#method-Option-check'><code>jmvcore::Option$check()</code></a></span></li>
<li><span class="pkg-link" data-pkg="jmvcore" data-topic="Option" data-id="getBoundValue"><a href='../../jmvcore/html/Option.html#method-Option-getBoundValue'><code>jmvcore::Option$getBoundValue()</code></a></span></li>
<li><span class="pkg-link" data-pkg="jmvcore" data-topic="OptionBool" data-id="initialize"><a href='../../jmvcore/html/OptionBool.html#method-OptionBool-initialize'><code>jmvcore::OptionBool$initialize()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OptionAction-clone"></a>}}
\if{latex}{\out{\hypertarget{method-OptionAction-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OptionAction$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
\section{Super class}{
\code{jmvcore::Option} -> \code{OptionList}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-OptionList-new}{\code{OptionList$new()}}
\item \href{#method-OptionList-clone}{\code{OptionList$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="jmvcore" data-topic="Option" data-id=".getData"><a href='../../jmvcore/html/Option.html#method-Option-.getData'><code>jmvcore::Option$.getData()</code></a></span></li>
<li><span class="pkg-link" data-pkg="jmvcore" data-topic="Option" data-id=".setParent"><a href='../../jmvcore/html/Option.html#method-Option-.setParent'><code>jmvcore::Option$.setParent()</code></a></span></li>
<li><span class="pkg-link" data-pkg="jmvcore" data-topic="Option" data-id="check"><a href='../../jmvcore/html/Option.html#method-Option-check'><code>jmvcore::Option$check()</code></a></span></li>
<li><span class="pkg-link" data-pkg="jmvcore" data-topic="Option" data-id="getBoundValue"><a href='../../jmvcore/html/Option.html#method-Option-getBoundValue'><code>jmvcore::Option$getBoundValue()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OptionList-new"></a>}}
\if{latex}{\out{\hypertarget{method-OptionList-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OptionList$new(name, value, options, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OptionList-clone"></a>}}
\if{latex}{\out{\hypertarget{method-OptionList-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OptionList$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
\section{Super class}{
\code{jmvcore::Option} -> \code{OptionNMXList}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-OptionNMXList-new}{\code{OptionNMXList$new()}}
\item \href{#method-OptionNMXList-clone}{\code{OptionNMXList$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="jmvcore" data-topic="Option" data-id=".getData"><a href='../../jmvcore/html/Option.html#method-Option-.getData'><code>jmvcore::Option$.getData()</code></a></span></li>
<li><span class="pkg-link" data-pkg="jmvcore" data-topic="Option" data-id=".setParent"><a href='../../jmvcore/html/Option.html#method-Option-.setParent'><code>jmvcore::Option$.setParent()</code></a></span></li>
<li><span class="pkg-link" data-pkg="jmvcore" data-topic="Option" data-id="check"><a href='../../jmvcore/html/Option.html#method-Option-check'><code>jmvcore::Option$check()</code></a></span></li>
<li><span class="pkg-link" data-pkg="jmvcore" data-topic="Option" data-id="getBoundValue"><a href='../../jmvcore/html/Option.html#method-Option-getBoundValue'><code>jmvcore::Option$getBoundValue()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OptionNMXList-new"></a>}}
\if{latex}{\out{\hypertarget{method-OptionNMXList-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OptionNMXList$new(name, value = character(), options, default = NULL, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OptionNMXList-clone"></a>}}
\if{latex}{\out{\hypertarget{method-OptionNMXList-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OptionNMXList$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
\section{Super class}{
\code{jmvcore::Option} -> \code{OptionVariables}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-OptionVariables-clone}{\code{OptionVariables$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="jmvcore" data-topic="Option" data-id=".getData"><a href='../../jmvcore/html/Option.html#method-Option-.getData'><code>jmvcore::Option$.getData()</code></a></span></li>
<li><span class="pkg-link" data-pkg="jmvcore" data-topic="Option" data-id=".setParent"><a href='../../jmvcore/html/Option.html#method-Option-.setParent'><code>jmvcore::Option$.setParent()</code></a></span></li>
<li><span class="pkg-link" data-pkg="jmvcore" data-topic="Option" data-id="check"><a href='../../jmvcore/html/Option.html#method-Option-check'><code>jmvcore::Option$check()</code></a></span></li>
<li><span class="pkg-link" data-pkg="jmvcore" data-topic="Option" data-id="getBoundValue"><a href='../../jmvcore/html/Option.html#method-Option-getBoundValue'><code>jmvcore::Option$getBoundValue()</code></a></span></li>
<li><span class="pkg-link" data-pkg="jmvcore" data-topic="Option" data-id="initialize"><a href='../../jmvcore/html/Option.html#method-Option-initialize'><code>jmvcore::Option$initialize()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OptionVariables-clone"></a>}}
\if{latex}{\out{\hypertarget{method-OptionVariables-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OptionVariables$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
\section{Super classes}{
\code{jmvcore::Option} -> \code{\link[jmvcore:OptionVariables]{jmvcore::OptionVariables}} -> \code{OptionVariables}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-OptionVariables-clone}{\code{OptionTerm$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="jmvcore" data-topic="Option" data-id=".getData"><a href='../../jmvcore/html/Option.html#method-Option-.getData'><code>jmvcore::Option$.getData()</code></a></span></li>
<li><span class="pkg-link" data-pkg="jmvcore" data-topic="Option" data-id=".setParent"><a href='../../jmvcore/html/Option.html#method-Option-.setParent'><code>jmvcore::Option$.setParent()</code></a></span></li>
<li><span class="pkg-link" data-pkg="jmvcore" data-topic="Option" data-id="check"><a href='../../jmvcore/html/Option.html#method-Option-check'><code>jmvcore::Option$check()</code></a></span></li>
<li><span class="pkg-link" data-pkg="jmvcore" data-topic="Option" data-id="getBoundValue"><a href='../../jmvcore/html/Option.html#method-Option-getBoundValue'><code>jmvcore::Option$getBoundValue()</code></a></span></li>
<li><span class="pkg-link" data-pkg="jmvcore" data-topic="Option" data-id="initialize"><a href='../../jmvcore/html/Option.html#method-Option-initialize'><code>jmvcore::Option$initialize()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OptionVariables-clone"></a>}}
\if{latex}{\out{\hypertarget{method-OptionVariables-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OptionTerm$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
\section{Super classes}{
\code{jmvcore::Option} -> \code{\link[jmvcore:OptionString]{jmvcore::OptionString}} -> \code{OptionVariable}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-OptionVariable-clone}{\code{OptionVariable$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="jmvcore" data-topic="Option" data-id=".getData"><a href='../../jmvcore/html/Option.html#method-Option-.getData'><code>jmvcore::Option$.getData()</code></a></span></li>
<li><span class="pkg-link" data-pkg="jmvcore" data-topic="Option" data-id=".setParent"><a href='../../jmvcore/html/Option.html#method-Option-.setParent'><code>jmvcore::Option$.setParent()</code></a></span></li>
<li><span class="pkg-link" data-pkg="jmvcore" data-topic="Option" data-id="check"><a href='../../jmvcore/html/Option.html#method-Option-check'><code>jmvcore::Option$check()</code></a></span></li>
<li><span class="pkg-link" data-pkg="jmvcore" data-topic="Option" data-id="getBoundValue"><a href='../../jmvcore/html/Option.html#method-Option-getBoundValue'><code>jmvcore::Option$getBoundValue()</code></a></span></li>
<li><span class="pkg-link" data-pkg="jmvcore" data-topic="Option" data-id="initialize"><a href='../../jmvcore/html/Option.html#method-Option-initialize'><code>jmvcore::Option$initialize()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OptionVariable-clone"></a>}}
\if{latex}{\out{\hypertarget{method-OptionVariable-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OptionVariable$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
\section{Super class}{
\code{jmvcore::Option} -> \code{OptionOutput}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-OptionOutput-clone}{\code{OptionOutput$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="jmvcore" data-topic="Option" data-id=".getData"><a href='../../jmvcore/html/Option.html#method-Option-.getData'><code>jmvcore::Option$.getData()</code></a></span></li>
<li><span class="pkg-link" data-pkg="jmvcore" data-topic="Option" data-id=".setParent"><a href='../../jmvcore/html/Option.html#method-Option-.setParent'><code>jmvcore::Option$.setParent()</code></a></span></li>
<li><span class="pkg-link" data-pkg="jmvcore" data-topic="Option" data-id="check"><a href='../../jmvcore/html/Option.html#method-Option-check'><code>jmvcore::Option$check()</code></a></span></li>
<li><span class="pkg-link" data-pkg="jmvcore" data-topic="Option" data-id="getBoundValue"><a href='../../jmvcore/html/Option.html#method-Option-getBoundValue'><code>jmvcore::Option$getBoundValue()</code></a></span></li>
<li><span class="pkg-link" data-pkg="jmvcore" data-topic="Option" data-id="initialize"><a href='../../jmvcore/html/Option.html#method-Option-initialize'><code>jmvcore::Option$initialize()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OptionOutput-clone"></a>}}
\if{latex}{\out{\hypertarget{method-OptionOutput-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OptionOutput$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
\section{Super classes}{
\code{jmvcore::Option} -> \code{\link[jmvcore:OptionArray]{jmvcore::OptionArray}} -> \code{OptionTerms}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-OptionTerms-new}{\code{OptionTerms$new()}}
\item \href{#method-OptionTerms-clone}{\code{OptionTerms$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="jmvcore" data-topic="Option" data-id=".getData"><a href='../../jmvcore/html/Option.html#method-Option-.getData'><code>jmvcore::Option$.getData()</code></a></span></li>
<li><span class="pkg-link" data-pkg="jmvcore" data-topic="Option" data-id=".setParent"><a href='../../jmvcore/html/Option.html#method-Option-.setParent'><code>jmvcore::Option$.setParent()</code></a></span></li>
<li><span class="pkg-link" data-pkg="jmvcore" data-topic="Option" data-id="check"><a href='../../jmvcore/html/Option.html#method-Option-check'><code>jmvcore::Option$check()</code></a></span></li>
<li><span class="pkg-link" data-pkg="jmvcore" data-topic="Option" data-id="getBoundValue"><a href='../../jmvcore/html/Option.html#method-Option-getBoundValue'><code>jmvcore::Option$getBoundValue()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OptionTerms-new"></a>}}
\if{latex}{\out{\hypertarget{method-OptionTerms-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OptionTerms$new(name, value, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OptionTerms-clone"></a>}}
\if{latex}{\out{\hypertarget{method-OptionTerms-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OptionTerms$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
\section{Super class}{
\code{jmvcore::Option} -> \code{OptionInteger}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-OptionInteger-clone}{\code{OptionInteger$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="jmvcore" data-topic="Option" data-id=".getData"><a href='../../jmvcore/html/Option.html#method-Option-.getData'><code>jmvcore::Option$.getData()</code></a></span></li>
<li><span class="pkg-link" data-pkg="jmvcore" data-topic="Option" data-id=".setParent"><a href='../../jmvcore/html/Option.html#method-Option-.setParent'><code>jmvcore::Option$.setParent()</code></a></span></li>
<li><span class="pkg-link" data-pkg="jmvcore" data-topic="Option" data-id="check"><a href='../../jmvcore/html/Option.html#method-Option-check'><code>jmvcore::Option$check()</code></a></span></li>
<li><span class="pkg-link" data-pkg="jmvcore" data-topic="Option" data-id="getBoundValue"><a href='../../jmvcore/html/Option.html#method-Option-getBoundValue'><code>jmvcore::Option$getBoundValue()</code></a></span></li>
<li><span class="pkg-link" data-pkg="jmvcore" data-topic="Option" data-id="initialize"><a href='../../jmvcore/html/Option.html#method-Option-initialize'><code>jmvcore::Option$initialize()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OptionInteger-clone"></a>}}
\if{latex}{\out{\hypertarget{method-OptionInteger-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OptionInteger$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
\section{Super class}{
\code{jmvcore::Option} -> \code{OptionNumber}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-OptionNumber-new}{\code{OptionNumber$new()}}
\item \href{#method-OptionNumber-clone}{\code{OptionNumber$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="jmvcore" data-topic="Option" data-id=".getData"><a href='../../jmvcore/html/Option.html#method-Option-.getData'><code>jmvcore::Option$.getData()</code></a></span></li>
<li><span class="pkg-link" data-pkg="jmvcore" data-topic="Option" data-id=".setParent"><a href='../../jmvcore/html/Option.html#method-Option-.setParent'><code>jmvcore::Option$.setParent()</code></a></span></li>
<li><span class="pkg-link" data-pkg="jmvcore" data-topic="Option" data-id="check"><a href='../../jmvcore/html/Option.html#method-Option-check'><code>jmvcore::Option$check()</code></a></span></li>
<li><span class="pkg-link" data-pkg="jmvcore" data-topic="Option" data-id="getBoundValue"><a href='../../jmvcore/html/Option.html#method-Option-getBoundValue'><code>jmvcore::Option$getBoundValue()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OptionNumber-new"></a>}}
\if{latex}{\out{\hypertarget{method-OptionNumber-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OptionNumber$new(name, value = 0, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OptionNumber-clone"></a>}}
\if{latex}{\out{\hypertarget{method-OptionNumber-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OptionNumber$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
\section{Super class}{
\code{jmvcore::Option} -> \code{OptionString}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-OptionString-clone}{\code{OptionString$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="jmvcore" data-topic="Option" data-id=".getData"><a href='../../jmvcore/html/Option.html#method-Option-.getData'><code>jmvcore::Option$.getData()</code></a></span></li>
<li><span class="pkg-link" data-pkg="jmvcore" data-topic="Option" data-id=".setParent"><a href='../../jmvcore/html/Option.html#method-Option-.setParent'><code>jmvcore::Option$.setParent()</code></a></span></li>
<li><span class="pkg-link" data-pkg="jmvcore" data-topic="Option" data-id="check"><a href='../../jmvcore/html/Option.html#method-Option-check'><code>jmvcore::Option$check()</code></a></span></li>
<li><span class="pkg-link" data-pkg="jmvcore" data-topic="Option" data-id="getBoundValue"><a href='../../jmvcore/html/Option.html#method-Option-getBoundValue'><code>jmvcore::Option$getBoundValue()</code></a></span></li>
<li><span class="pkg-link" data-pkg="jmvcore" data-topic="Option" data-id="initialize"><a href='../../jmvcore/html/Option.html#method-Option-initialize'><code>jmvcore::Option$initialize()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OptionString-clone"></a>}}
\if{latex}{\out{\hypertarget{method-OptionString-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OptionString$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
\section{Super class}{
\code{jmvcore::Option} -> \code{OptionString}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-OptionString-clone}{\code{OptionLevel$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="jmvcore" data-topic="Option" data-id=".getData"><a href='../../jmvcore/html/Option.html#method-Option-.getData'><code>jmvcore::Option$.getData()</code></a></span></li>
<li><span class="pkg-link" data-pkg="jmvcore" data-topic="Option" data-id=".setParent"><a href='../../jmvcore/html/Option.html#method-Option-.setParent'><code>jmvcore::Option$.setParent()</code></a></span></li>
<li><span class="pkg-link" data-pkg="jmvcore" data-topic="Option" data-id="check"><a href='../../jmvcore/html/Option.html#method-Option-check'><code>jmvcore::Option$check()</code></a></span></li>
<li><span class="pkg-link" data-pkg="jmvcore" data-topic="Option" data-id="getBoundValue"><a href='../../jmvcore/html/Option.html#method-Option-getBoundValue'><code>jmvcore::Option$getBoundValue()</code></a></span></li>
<li><span class="pkg-link" data-pkg="jmvcore" data-topic="Option" data-id="initialize"><a href='../../jmvcore/html/Option.html#method-Option-initialize'><code>jmvcore::Option$initialize()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OptionString-clone"></a>}}
\if{latex}{\out{\hypertarget{method-OptionString-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OptionLevel$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
\section{Super class}{
\code{jmvcore::Option} -> \code{OptionGroup}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-OptionGroup-new}{\code{OptionGroup$new()}}
\item \href{#method-OptionGroup-clone}{\code{OptionGroup$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="jmvcore" data-topic="Option" data-id=".getData"><a href='../../jmvcore/html/Option.html#method-Option-.getData'><code>jmvcore::Option$.getData()</code></a></span></li>
<li><span class="pkg-link" data-pkg="jmvcore" data-topic="Option" data-id=".setParent"><a href='../../jmvcore/html/Option.html#method-Option-.setParent'><code>jmvcore::Option$.setParent()</code></a></span></li>
<li><span class="pkg-link" data-pkg="jmvcore" data-topic="Option" data-id="check"><a href='../../jmvcore/html/Option.html#method-Option-check'><code>jmvcore::Option$check()</code></a></span></li>
<li><span class="pkg-link" data-pkg="jmvcore" data-topic="Option" data-id="getBoundValue"><a href='../../jmvcore/html/Option.html#method-Option-getBoundValue'><code>jmvcore::Option$getBoundValue()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OptionGroup-new"></a>}}
\if{latex}{\out{\hypertarget{method-OptionGroup-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OptionGroup$new(name, value, elements, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OptionGroup-clone"></a>}}
\if{latex}{\out{\hypertarget{method-OptionGroup-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OptionGroup$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
\section{Super classes}{
\code{jmvcore::Option} -> \code{\link[jmvcore:OptionGroup]{jmvcore::OptionGroup}} -> \code{OptionPair}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-OptionPair-new}{\code{OptionPair$new()}}
\item \href{#method-OptionPair-clone}{\code{OptionPair$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="jmvcore" data-topic="Option" data-id=".getData"><a href='../../jmvcore/html/Option.html#method-Option-.getData'><code>jmvcore::Option$.getData()</code></a></span></li>
<li><span class="pkg-link" data-pkg="jmvcore" data-topic="Option" data-id=".setParent"><a href='../../jmvcore/html/Option.html#method-Option-.setParent'><code>jmvcore::Option$.setParent()</code></a></span></li>
<li><span class="pkg-link" data-pkg="jmvcore" data-topic="Option" data-id="check"><a href='../../jmvcore/html/Option.html#method-Option-check'><code>jmvcore::Option$check()</code></a></span></li>
<li><span class="pkg-link" data-pkg="jmvcore" data-topic="Option" data-id="getBoundValue"><a href='../../jmvcore/html/Option.html#method-Option-getBoundValue'><code>jmvcore::Option$getBoundValue()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OptionPair-new"></a>}}
\if{latex}{\out{\hypertarget{method-OptionPair-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OptionPair$new(name, value, permitted = NULL, suggested = NULL, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OptionPair-clone"></a>}}
\if{latex}{\out{\hypertarget{method-OptionPair-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OptionPair$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
\section{Super classes}{
\code{jmvcore::Option} -> \code{\link[jmvcore:OptionGroup]{jmvcore::OptionGroup}} -> \code{OptionSort}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-OptionSort-new}{\code{OptionSort$new()}}
\item \href{#method-OptionSort-clone}{\code{OptionSort$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="jmvcore" data-topic="Option" data-id=".getData"><a href='../../jmvcore/html/Option.html#method-Option-.getData'><code>jmvcore::Option$.getData()</code></a></span></li>
<li><span class="pkg-link" data-pkg="jmvcore" data-topic="Option" data-id=".setParent"><a href='../../jmvcore/html/Option.html#method-Option-.setParent'><code>jmvcore::Option$.setParent()</code></a></span></li>
<li><span class="pkg-link" data-pkg="jmvcore" data-topic="Option" data-id="check"><a href='../../jmvcore/html/Option.html#method-Option-check'><code>jmvcore::Option$check()</code></a></span></li>
<li><span class="pkg-link" data-pkg="jmvcore" data-topic="Option" data-id="getBoundValue"><a href='../../jmvcore/html/Option.html#method-Option-getBoundValue'><code>jmvcore::Option$getBoundValue()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OptionSort-new"></a>}}
\if{latex}{\out{\hypertarget{method-OptionSort-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OptionSort$new(name, value, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OptionSort-clone"></a>}}
\if{latex}{\out{\hypertarget{method-OptionSort-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OptionSort$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
\section{Super class}{
\code{jmvcore::Option} -> \code{OptionArray}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-OptionArray-new}{\code{OptionArray$new()}}
\item \href{#method-OptionArray-clone}{\code{OptionArray$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="jmvcore" data-topic="Option" data-id=".getData"><a href='../../jmvcore/html/Option.html#method-Option-.getData'><code>jmvcore::Option$.getData()</code></a></span></li>
<li><span class="pkg-link" data-pkg="jmvcore" data-topic="Option" data-id=".setParent"><a href='../../jmvcore/html/Option.html#method-Option-.setParent'><code>jmvcore::Option$.setParent()</code></a></span></li>
<li><span class="pkg-link" data-pkg="jmvcore" data-topic="Option" data-id="check"><a href='../../jmvcore/html/Option.html#method-Option-check'><code>jmvcore::Option$check()</code></a></span></li>
<li><span class="pkg-link" data-pkg="jmvcore" data-topic="Option" data-id="getBoundValue"><a href='../../jmvcore/html/Option.html#method-Option-getBoundValue'><code>jmvcore::Option$getBoundValue()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OptionArray-new"></a>}}
\if{latex}{\out{\hypertarget{method-OptionArray-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OptionArray$new(name, value, template, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OptionArray-clone"></a>}}
\if{latex}{\out{\hypertarget{method-OptionArray-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OptionArray$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
\section{Super classes}{
\code{jmvcore::Option} -> \code{\link[jmvcore:OptionArray]{jmvcore::OptionArray}} -> \code{OptionPairs}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-OptionPairs-new}{\code{OptionPairs$new()}}
\item \href{#method-OptionPairs-clone}{\code{OptionPairs$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="jmvcore" data-topic="Option" data-id=".getData"><a href='../../jmvcore/html/Option.html#method-Option-.getData'><code>jmvcore::Option$.getData()</code></a></span></li>
<li><span class="pkg-link" data-pkg="jmvcore" data-topic="Option" data-id=".setParent"><a href='../../jmvcore/html/Option.html#method-Option-.setParent'><code>jmvcore::Option$.setParent()</code></a></span></li>
<li><span class="pkg-link" data-pkg="jmvcore" data-topic="Option" data-id="check"><a href='../../jmvcore/html/Option.html#method-Option-check'><code>jmvcore::Option$check()</code></a></span></li>
<li><span class="pkg-link" data-pkg="jmvcore" data-topic="Option" data-id="getBoundValue"><a href='../../jmvcore/html/Option.html#method-Option-getBoundValue'><code>jmvcore::Option$getBoundValue()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OptionPairs-new"></a>}}
\if{latex}{\out{\hypertarget{method-OptionPairs-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OptionPairs$new(name, value, permitted = NULL, suggested = NULL, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OptionPairs-clone"></a>}}
\if{latex}{\out{\hypertarget{method-OptionPairs-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OptionPairs$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
