% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\docType{data}
\name{Options}
\alias{Options}
\alias{OptionBool}
\alias{OptionAction}
\alias{OptionList}
\alias{OptionNMXList}
\alias{OptionVariables}
\alias{OptionTerm}
\alias{OptionVariable}
\alias{OptionOutput}
\alias{OptionTerms}
\alias{OptionInteger}
\alias{OptionNumber}
\alias{OptionString}
\alias{OptionLevel}
\alias{OptionGroup}
\alias{OptionPair}
\alias{OptionSort}
\alias{OptionArray}
\alias{OptionPairs}
\title{The jmv Options classes}
\format{An object of class \code{R6ClassGenerator} of length 25.}
\usage{
Options

OptionBool

OptionAction

OptionList

OptionNMXList

OptionVariables

OptionTerm

OptionVariable

OptionOutput

OptionTerms

OptionInteger

OptionNumber

OptionString

OptionLevel

OptionGroup

OptionPair

OptionSort

OptionArray

OptionPairs
}
\description{
The jmv Options classes
}
\keyword{datasets}
