% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\docType{data}
\name{Options}
\alias{Options}
\alias{OptionBool}
\alias{OptionList}
\alias{OptionNMXList}
\alias{OptionVariables}
\alias{OptionVariable}
\alias{OptionTerms}
\alias{OptionInteger}
\alias{OptionNumber}
\alias{OptionString}
\alias{OptionGroup}
\alias{OptionSort}
\alias{OptionArray}
\alias{OptionPairs}
\title{The jmv Options classes}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
Options

OptionBool

OptionList

OptionNMXList

OptionVariables

OptionVariable

OptionTerms

OptionInteger

OptionNumber

OptionString

OptionGroup

OptionSort

OptionArray

OptionPairs
}
\description{
The jmv Options classes
}
\keyword{datasets}
