% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themes.R
\name{theme_hadley}
\alias{theme_hadley}
\title{Creates the hadley jmv ggplot2 theme}
\usage{
theme_hadley(base_size = 16, scale = "none", palette = "jmv")
}
\arguments{
\item{base_size}{Font size}

\item{scale}{'none' or 'discrete'}

\item{palette}{Color palette name}
}
\value{
the hadley jmv ggplot2 theme
}
\description{
Creates the hadley jmv ggplot2 theme
}
